requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow
anthropic
httpx
python-multipart

# dev / test
pytest
pytest-asyncio
